backend/ifconfig/views.py
# -*- coding: utf-8 -*-

"""
 *  Copyright (C) 2011-2016, it-novum GmbH <community@openattic.org>
 *
 *  openATTIC is free software; you can redistribute it and/or modify it
 *  under the terms of the GNU General Public License as published by
 *  the Free Software Foundation; version 2.
 *
 *  This package is distributed in the hope that it will be useful,
 *  but WITHOUT ANY WARRANTY; without even the implied warranty of
 *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *  GNU General Public License for more details.
"""

from django.conf import settings

from rest_framework.decorators import api_view
from rest_framework.response import Response

@api_view(['GET'])
def current_host(request):
    res = {'oa_version': {'package': {'VERSION': settings.VERSION_CONF['package']['VERSION']}}}
    return Response(res)



