requirements/ubuntu-16.04.txt
decorator==4.0.9
Django==1.8.7
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.3.2
djangorestframework-bulk==0.2.1
ipython==4.2.0
ipython-genutils==0.1.0
mock==2.0.0
pyinotify==0.9.6
requests==2.9.1
six==1.10.0
docopt==0.6.2
configobj
requests-aws==0.1.8
testtools==2.3.0
xmlrunner==1.7.7
